Refactoring - prepare for move PartitionChooser to PartitionWriter
