+++ ydb/services/persqueue_v1/actors/partition_writer_cache_actor.cpp
@@ -6,14 +6,10 @@ namespace NKikimr::NGRpcProxy::V1 {
 TPartitionWriterCacheActor::TPartitionWriterCacheActor(const TActorId& owner,
                                                        ui32 partition,
                                                        ui64 tabletId,
-                                                       std::optional<ui32> expectedGeneration,
-                                                       const TString& sourceId,
                                                        const NPQ::TPartitionWriterOpts& opts) :
     Owner(owner),
     Partition(partition),
     TabletId(tabletId),
-    ExpectedGeneration(expectedGeneration),
-    SourceId(sourceId),
     Opts(opts)
 {
 }
@@ -50,7 +46,7 @@ STFUNC(TPartitionWriterCacheActor::StateWork)
         HFunc(NPQ::TEvPartitionWriter::TEvWriteAccepted, Handle);
         HFunc(NPQ::TEvPartitionWriter::TEvWriteResponse, Handle);
         HFunc(NPQ::TEvPartitionWriter::TEvDisconnected, Handle);
-        HFunc(TEvents::TEvPoisonPill, Handle);
+        HFunc(TEvents::TEvPoison, Handle);
     }
 }
 
@@ -270,8 +266,7 @@ TActorId TPartitionWriterCacheActor::CreatePartitionWriter(const TString& sessio
     }
 
     return ctx.RegisterWithSameMailbox(NPQ::CreatePartitionWriter(
-        ctx.SelfID, {}, TabletId, Partition, ExpectedGeneration,
-        SourceId, opts
+        ctx.SelfID, TabletId, Partition, opts
     ));
 }
 
@@ -283,7 +278,7 @@ STFUNC(TPartitionWriterCacheActor::StateBroken)
         HFunc(NPQ::TEvPartitionWriter::TEvWriteAccepted, Handle);
         HFunc(NPQ::TEvPartitionWriter::TEvWriteResponse, Handle);
         HFunc(NPQ::TEvPartitionWriter::TEvDisconnected, Handle);
-        HFunc(TEvents::TEvPoisonPill, Handle);
+        HFunc(TEvents::TEvPoison, Handle);
     }
 }
